# api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is the API for my personal dev website",
		"title": "Dev Website API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"name": "Matheo Galuba",
			"url": "https://dev.matheo-galuba.com/contact",
			"email": "matheo.galu56@gmail.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		},
		"version": "1.0"
	},
	"host": "localhost:8080",
	"basePath": "/api",
	"paths": {
...
		"/auth/passkey/begin": {
			"post": {
				"description": "Get the options to pass to navigator.credentials.get(). No email is needed, the passkey identifies the user.",
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Start a passkey login",
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/controllers.PasskeyBeginResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/auth/passkey/finish": {
			"post": {
				"description": "Verify the assertion returned by navigator.credentials.get() and log the owner of the passkey in",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["auth"],
				"summary": "Finish a passkey login",
				"parameters": [
					{
						"description": "Session and assertion",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.PasskeyFinishBody"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": true
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
...
		"/users/{id}/passkeys/begin": {
			"post": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Get the options to pass to navigator.credentials.create() to add a passkey to the authenticated user",
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Start adding a passkey",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/controllers.PasskeyBeginResponse"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/users/{id}/passkeys/finish": {
			"post": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Verify the credential returned by navigator.credentials.create() and add it to the authenticated user",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Finish adding a passkey",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"description": "Session, passkey name and credential",
						"name": "body",
						"in": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/controllers.PasskeyFinishBody"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/models.User"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/users/{id}/passkeys/{passkeyId}": {
			"delete": {
				"security": [
					{
						"Bearer": []
					}
				],
				"description": "Remove a passkey from the authenticated user",
				"produces": ["application/json"],
				"tags": ["user"],
				"summary": "Remove a passkey",
				"parameters": [
					{
						"type": "string",
						"description": "User ID",
						"name": "id",
						"in": "path",
						"required": true
					},
					{
						"type": "string",
						"description": "Passkey ID",
						"name": "passkeyId",
						"in": "path",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/models.User"
						}
					},
					"400": {
						"description": "Bad Request",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "string"
							}
						}
					}
				}
			}
		}
	},
	"definitions": {
...
		"controllers.PasskeyBeginResponse": {
			"type": "object",
			"properties": {
				"options": {},
				"sessionId": {
					"type": "string"
				}
			}
		},
		"controllers.PasskeyFinishBody": {
			"type": "object",
			"required": ["credential", "sessionId"],
			"properties": {
				"credential": {
					"type": "object"
				},
				"name": {
					"type": "string",
					"maxLength": 50
				},
				"sessionId": {
					"type": "string"
				}
			}
		},
...
		"models.Identity": {
			"type": "object",
			"properties": {
				"avatarUrl": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"lastLogin": {
					"type": "string"
				},
				"linkedAt": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"profileUrl": {
					"type": "string"
				},
				"provider": {
					"type": "string",
					"enum": ["email", "github", "google"]
				},
				"providerId": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			}
		},
...
		"models.Passkey": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"lastUsed": {
					"type": "string"
				},
				"name": {
					"type": "string"
				}
			}
		},
...
		"models.User": {
			"type": "object",
			"properties": {
				"createdAt": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"flavour": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"identities": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.Identity"
					}
				},
				"lastLogin": {
					"type": "string"
				},
				"lastRefresh": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"passkeys": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/models.Passkey"
					}
				},
				"profilePicture": {
					"type": "string"
				}
			}
		}
	},
	"securityDefinitions": {
		"Bearer": {
			"description": "Bearer token",
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}